requirements.txt
wheel==0.37.0
certifi==2021.5.30
charset-normalizer==2.0.4
idna==3.2
mwparserfromhell==0.6.3
pywikibot==6.5.0
requests==2.26.0
six==1.16.0
sseclient==0.0.27
urllib3==1.26.6
python-dotenv==0.19.0
PyMySQL==1.0.2
discord.py==1.7.3